Extract tracking tab
